Extract setup prompt tests
Move the prompt helper tests into a sibling test module so the\nproduction file no longer carries the inline test block. This keeps\nthe behaviour unchanged while preserving the validated setup::prompts\ntest coverage in the new file-backed test module.
